backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic-settings>=2.4.0,<3.0.0
groq>=0.11.0,<1.0.0
supabase>=2.6.0,<3.0.0
PyMuPDF>=1.24.0,<2.0.0
fpdf2>=2.8.1,<3.0.0
httpx>=0.27.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
email-validator>=2.2.0,<3.0.0
pytest>=8.3.0,<9.0.0

